+++ lib/src/labels/label.dart
@@ -13,7 +13,7 @@ class Label extends StatelessWidget {
     this.icon,
     required this.text,
     this.child,
-    this.yAlignment = CrossAxisAlignment.start,
+    this.crossAxisAlignment = CrossAxisAlignment.start,
   });
 
   /// The icon used by the label. If non-null, it's rendered horizontally
@@ -29,7 +29,8 @@ class Label extends StatelessWidget {
   /// The widget at the right of [text].
   final Widget? child;
 
-  final CrossAxisAlignment yAlignment;
+  /// The cross-axis alignment of the label.
+  final CrossAxisAlignment crossAxisAlignment;
 
   @override
   Widget build(BuildContext context) {
@@ -43,7 +44,7 @@ class Label extends StatelessWidget {
       child: this.text,
     );
     return Row(
-      crossAxisAlignment: yAlignment,
+      crossAxisAlignment: crossAxisAlignment,
       mainAxisSize: MainAxisSize.min,
       children: [
         if (icon != null)