+++ pr_prelaunch_tasks.sh
@@ -4,18 +4,18 @@ if [ $? -eq 1 ]; then
   git add .
   git commit -m "chore: run flutter format ."
   echo "push changes? [y/n]"
-  read pushResponse
-  if [ $pushResponse = "y" ]; then
+  read -r pushResponse
+  if [ "$pushResponse" = "y" ]; then
     git push origin
   fi
 fi
 echo "Run dart fix --dry-run? [y/n]"
-read dryRunResponse
+read -r dryRunResponse
 if [ "$dryRunResponse" = "y" ]; then
   dart fix --dry-run
 fi
 echo "Run dart fix --apply? [y/n]"
-read applyResponse
+read -r applyResponse
 if [ "$applyResponse" = "y" ]; then
   dart fix --apply
   if [ -z "$(git status --porcelain)" ]; then
@@ -24,14 +24,14 @@ if [ "$applyResponse" = "y" ]; then
     git add .
     git commit -m "chore: run dart fix --apply"
     echo "push changes? [y/n]"
-    read pushResponse
-    if [ $pushResponse = "y" ]; then
+    read -r pushResponse
+    if [ "$pushResponse" = "y" ]; then
       git push origin
     fi
   fi
 fi
 echo "Run tests? [y/n]"
-read testResponse
+read -r testResponse
 if [ "$testResponse" = "y" ]; then
   flutter test
 else